Basic Shader/Model Testing
Early testing with the URP for character outlining with the new Mercy default model. This isn't smoothly compatibly with the pixel-perfect camera, so it'll need a method to make it conform nicely to that. Later additions like lighting will also add conflicts so it'll need a resolution by then.

Also created a new 320x180 testing map. Movement still isn't implemented, but will be soon.